+++ docs/release_notes.rst
@@ -17,6 +17,7 @@ New Features
 * Added support for Python 3.6 (#270).
 * Specifying an invalid error code prefix (e.g., ``--select=D9``) will print
   a warning message to ``stderr`` (#253, #279).
+* Configuration files now support multiple-lined entries (#250, #281).
 
 Bug Fixes